Draft

+++ csharp/Platform.Data.Doublets.Sequences.Tests/ByteConvertersTests.cs
@@ -18,7 +18,7 @@ using Platform.Memory;
 using Platform.Numbers;
 using Platform.Reflection;
 using Xunit;
-using TLinkAddress = System.UInt32;
+using TLinkAddress = System.UInt16;
 
 namespace Platform.Data.Doublets.Sequences.Tests
 {
